Update the project

+++ service-order/service-order-ejb/src/main/java/br/com/codeshare/service/ClientService.java
@@ -25,7 +25,9 @@ public class ClientService{
 	private ServiceOrderService soService;
 	
 	public void save(Client client) throws Exception{
-		
+
+		validatePhoneLeastOnePhoneObligatory(client);
+
 		clientRepository.insert(client);
 		clientEventSrc.fire(client);
 	}
@@ -44,13 +46,17 @@ public class ClientService{
 	}
 
 	public void update(Client client,List<Phone>phonesToBeRemove) throws Exception {
-		
+
+		validatePhoneLeastOnePhoneObligatory(client);
+
 		clientRepository.update(client);
 		
 		if(phonesToBeRemove != null){
 			for(Phone phone : phonesToBeRemove){
 				if(soService.findSoByPhoneId(phone.getId()).isEmpty()){
-					phoneService.remove(phone);
+					if(phone.getId()!=null){
+						phoneService.remove(phone);
+					}
 				}
 				else{
 					throw new BusinessException(ErrorCode.PHONE_HAS_SERVICE_ORDER.getErrorCode());
@@ -63,6 +69,12 @@ public class ClientService{
 	
 	public void removePhoneCliente(Client client, Phone phone){
 		clientRepository.removePhoneClient(client, phone);
+	}
+
+	private void validatePhoneLeastOnePhoneObligatory(Client client) throws BusinessException {
+		if(client.getHomePhone().isEmpty() && client.getBisenessPhone().isEmpty()){
+			throw new BusinessException(ErrorCode.LEAST_ONE_PHONE_OBLIGATORY.getErrorCode());
+		}
 	}
 	
 }
